core/requirements.txt
gevent==1.1.2
Flask==0.10.1
SQLAlchemy==1.0.11
alembic==0.8.3
redis==2.10.3
jsonschema==2.3.0
PyMySQL==0.6.7
warlock==1.0.1
python-glanceclient==2.0.0
python-keystoneclient==2.3.1
python-neutronclient==4.1.1
python-novaclient==3.3.1
python-ceilometerclient==2.3.0
